refactor and update todo
