+++ README.md
@@ -16,8 +16,6 @@ Snack Hunter: The Ultimate Taiwanese Night Market Food Finder
 
 [2024/12/26] Refactor the whole codebase (Frontend: APP, Web, Backend, Models, Docs)
 
-[2024/12/25] Update SearchViewModel.swift version
-
 [2024/12/25] Update this README.md: `Source Code Structure` and `Project introduction`
 
 ## Source Code Structure
